requirement.txt
scipy==1.7.0
pandas==1.3.1
numpy==1.20.3
pytorch